@@ -1,58 +1,53 @@
 {
   "name": "traccar",
-  "version": "5.12.1",
+  "version": "6.5.0",
   "type": "module",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.11.3",
-    "@emotion/styled": "^11.11.0",
-    "@fontsource/roboto": "^5.0.8",
+    "@emotion/cache": "^11.13.1",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
     "@mapbox/mapbox-gl-draw": "^1.4.3",
-    "@maplibre/maplibre-gl-geocoder": "^1.5.0",
-    "@mui/icons-material": "^5.15.4",
-    "@mui/lab": "^5.0.0-alpha.160",
-    "@mui/material": "^5.15.4",
-    "@mui/styles": "^5.15.4",
-    "@mui/x-charts": "^7.3.0",
-    "@reduxjs/toolkit": "^2.0.1",
+    "@mapbox/mapbox-gl-rtl-text": "^0.2.3",
+    "@maplibre/maplibre-gl-geocoder": "^1.7.0",
+    "@mui/icons-material": "^5.16.7",
+    "@mui/lab": "^5.0.0-alpha.173",
+    "@mui/material": "^5.16.7",
+    "@mui/styles": "^5.16.7",
+    "@reduxjs/toolkit": "^2.2.7",
     "@tmcw/togeojson": "^5.8.1",
-    "@turf/circle": "^6.5.0",
-    "@turf/turf": "^6.5.0",
-    "@vitejs/plugin-react": "^4.2.1",
-    "braces": "^3.0.3",
-    "dayjs": "^1.11.10",
-    "debug": "^4.3.5",
-    "ejs": "^3.1.10",
+    "@turf/circle": "^7.1.0",
+    "@turf/turf": "^7.1.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "dayjs": "^1.11.13",
     "events": "^3.3.0",
-    "framer-motion": "^11.11.17",
     "mapbox-gl": "^1.13.3",
-    "maplibre-gl": "^3.6.2",
-    "react": "^18.2.0",
+    "maplibre-gl": "^4.7.0",
+    "maplibre-google-maps": "^1.1.0",
+    "react": "^18.3.1",
     "react-country-flag": "3.1.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-draggable": "^4.4.6",
-    "react-export-table-to-excel": "^1.0.6",
     "react-mui-dropzone": "^4.0.7",
-    "react-redux": "^9.1.0",
-    "react-router-dom": "^6.21.2",
-    "react-spring-bottom-sheet": "^3.4.1",
-    "react-virtualized-auto-sizer": "^1.0.20",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.26.2",
+    "react-virtualized-auto-sizer": "^1.0.24",
     "react-window": "^1.8.10",
-    "recharts": "^2.10.4",
+    "recharts": "^2.12.7",
     "redux": "^5.0.1",
-    "vite": "^5.3.1",
-    "vite-plugin-pwa": "^0.17.4",
+    "stylis": "^4.3.4",
+    "stylis-plugin-rtl": "^2.1.1",
+    "vite": "^5.4.6",
+    "vite-plugin-pwa": "^0.20.5",
     "vite-plugin-svgr": "^4.2.0",
-    "wellknown": "^0.5.0",
-    "xlsx": "^0.18.5"
+    "wellknown": "^0.5.0"
   },
   "scripts": {
     "start": "vite",
     "build": "vite build",
     "generate-pwa-assets": "pwa-assets-generator --preset minimal public/logo.svg",
     "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "build:dist": "react-build-dist"
+    "lint:fix": "eslint --fix ."
   },
   "browserslist": {
     "production": [
@@ -66,12 +61,12 @@
     ]
   },
   "devDependencies": {
-    "@vite-pwa/assets-generator": "^0.2.3",
-    "eslint": "^8.56.0",
+    "@vite-pwa/assets-generator": "^0.2.6",
+    "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0"
+    "eslint-plugin-import": "^2.30.0",
+    "eslint-plugin-jsx-a11y": "^6.10.0",
+    "eslint-plugin-react": "^7.36.1",
+    "eslint-plugin-react-hooks": "^4.6.2"
   }
 }
