@@ -1,10 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-
-import '@fontsource/roboto/300.css';
-import '@fontsource/roboto/400.css';
-import '@fontsource/roboto/500.css';
-import '@fontsource/roboto/700.css';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
@@ -21,7 +15,8 @@ import AppThemeProvider from './AppThemeProvider';
 
 preloadImages();
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <ErrorBoundary>
     <Provider store={store}>
       <LocalizationProvider>
@@ -40,27 +35,4 @@ ReactDOM.render(
       </LocalizationProvider>
     </Provider>
   </ErrorBoundary>,
-  document.getElementById('root')
 );
-
-// const root = createRoot(document.getElementById('root'));
-// root.render(
-//   <ErrorBoundary>
-//     <Provider store={store}>
-//       <LocalizationProvider>
-//         <StyledEngineProvider injectFirst>
-//           <AppThemeProvider>
-//             <CssBaseline />
-//             <ServerProvider>
-//               <BrowserRouter>
-//                 <Navigation />
-//               </BrowserRouter>
-//               <ErrorHandler />
-//               <NativeInterface />
-//             </ServerProvider>
-//           </AppThemeProvider>
-//         </StyledEngineProvider>
-//       </LocalizationProvider>
-//     </Provider>
-//   </ErrorBoundary>,
-// );
