@@ -16,7 +16,6 @@ const useStyles = makeStyles((theme) => ({
     height: 'auto',
     marginTop: 0,
     marginBottom: 0,
-    color: theme.palette.error.main,
   },
 }));
 
@@ -43,7 +42,7 @@ const RemoveDialog = ({
       onClose={() => onResult(false)}
       message={t('sharedRemoveConfirm')}
       action={(
-        <Button size="small" className={classes.button} onClick={handleRemove}>
+        <Button size="small" className={classes.button} color="error" onClick={handleRemove}>
           {t('sharedRemove')}
         </Button>
       )}
