@@ -9,15 +9,14 @@ export default [
   'crane',
   'helicopter',
   'motorcycle',
-  'offroad',
   'person',
   'pickup',
   'plane',
   'ship',
   'tractor',
+  'trailer',
   'train',
   'tram',
-  'trolleybus',
   'truck',
   'van',
   'suv',
