@@ -1,6 +1,7 @@
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
 import relativeTime from 'dayjs/plugin/relativeTime';
+import localizedFormat from 'dayjs/plugin/localizedFormat';
 
 import {
   altitudeFromMeters,
@@ -12,44 +13,54 @@ import {
   volumeFromLiters,
   volumeUnitString,
 } from './converter';
-import {
-  prefixString
-} from './stringUtils';
+import { prefixString } from './stringUtils';
 
 dayjs.extend(duration);
 dayjs.extend(relativeTime);
+dayjs.extend(localizedFormat);
 
 export const formatBoolean = (value, t) => (value ? t('sharedYes') : t('sharedNo'));
 
 export const formatNumber = (value, precision = 1) => Number(value.toFixed(precision));
 
-export const formatPercentage = (value) => `${value}%`;
+export const formatPercentage = (value) => `${value.toFixed(0)}%`;
 
-export const formatTemperature = (value) => `${value}°C`;
+export const formatTemperature = (value) => `${value.toFixed(1)}°C`;
 
-export const formatVoltage = (value, t) => `${value} ${t('sharedVoltAbbreviation')}`;
+export const formatVoltage = (value, t) => `${value.toFixed(2)} ${t('sharedVoltAbbreviation')}`;
 
-export const formatConsumption = (value, t) => `${value} ${t('sharedLiterPerHourAbbreviation')}`;
+export const formatConsumption = (value, t) => `${value.toFixed(2)} ${t('sharedLiterPerHourAbbreviation')}`;
 
-export const formatTime = (value, format, hours12) => {
+export const formatTime = (value, format) => {
   if (value) {
-    const d = dayjs(value);
+    const d = dayjs(value).toDate();
+    const dateConfig = { year: 'numeric', month: '2-digit', day: '2-digit' };
+    const minuteConfig = { hour: '2-digit', minute: '2-digit' };
+    const secondConfig = { ...minuteConfig, second: '2-digit' };
     switch (format) {
       case 'date':
-        return d.format('YYYY-MM-DD');
+        return d.toLocaleDateString(undefined, dateConfig);
       case 'time':
-        return d.format(hours12 ? 'hh:mm:ss A' : 'HH:mm:ss');
+        return d.toLocaleTimeString(undefined, secondConfig);
       case 'minutes':
-        return d.format(hours12 ? 'YYYY-MM-DD hh:mm A' : 'YYYY-MM-DD HH:mm');
+        return d.toLocaleString(undefined, { ...dateConfig, ...minuteConfig });
       default:
-        return d.format(hours12 ? 'YYYY-MM-DD hh:mm:ss A' : 'YYYY-MM-DD HH:mm:ss');
+        return d.toLocaleString(undefined, { ...dateConfig, ...secondConfig });
     }
   }
   return '';
 };
 
 export const formatStatus = (value, t) => t(prefixString('deviceStatus', value));
-export const formatAlarm = (value, t) => (value ? t(prefixString('alarm', value)) : '');
+
+export const formatAlarm = (value, t) => {
+  if (value) {
+    return value.split(',')
+      .map((alarm) => t(prefixString('alarm', alarm)))
+      .join(', ');
+  }
+  return '';
+};
 
 export const formatCourse = (value) => {
   const courseValues = ['\u2191', '\u2197', '\u2192', '\u2198', '\u2193', '\u2199', '\u2190', '\u2196'];
@@ -72,8 +83,6 @@ export const formatNumericHours = (value, t) => {
   const hours = Math.floor(value / 3600000);
   const minutes = Math.floor((value % 3600000) / 60000);
   return `${hours} ${t('sharedHourAbbreviation')} ${minutes} ${t('sharedMinuteAbbreviation')}`;
-  //const hours = value / 3600000;
-  //return hours.toFixed(1);
 };
 
 export const formatCoordinate = (key, value, unit) => {
@@ -105,17 +114,13 @@ export const formatCoordinate = (key, value, unit) => {
   }
 };
 
-export const getStatusColor = (status, lastUpdate) => {
+export const getStatusColor = (status) => {
   switch (status) {
     case 'online':
       return 'success';
     case 'offline':
       return 'error';
     case 'unknown':
-      if (lastUpdate && dayjs().diff(dayjs(lastUpdate), 'day') > 20) {
-        return 'error';
-      }
-      return 'neutral';
     default:
       return 'neutral';
   }
@@ -132,6 +137,9 @@ export const getBatteryStatus = (batteryLevel) => {
 };
 
 export const formatNotificationTitle = (t, notification, includeId) => {
+  if (notification.description) {
+    return notification.description;
+  }
   let title = t(prefixString('event', notification.type));
   if (notification.type === 'alarm') {
     const alarmString = notification.attributes.alarms;
