@@ -2,30 +2,26 @@ import React from 'react';
 import { useTheme, useMediaQuery } from '@mui/material';
 import { useSelector } from 'react-redux';
 import { makeStyles } from '@mui/styles';
-// import Logo from '../resources/images/logo.svg?react';
-import Logo from '../resources/images/logoGPS.svg?react';
+import Logo from '../resources/images/logo.svg?react';
 
-
-
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles((theme) => ({
   image: {
     alignSelf: 'center',
     maxWidth: '240px',
     maxHeight: '120px',
     width: 'auto',
     height: 'auto',
+    margin: theme.spacing(2),
   },
 }));
 
 const LogoImage = ({ color }) => {
-
   const theme = useTheme();
   const classes = useStyles();
 
   const expanded = !useMediaQuery(theme.breakpoints.down('lg'));
 
-  // const logo = useSelector((state) => state.session.server.attributes?.logo);
-  const logo = useSelector((state) => state.session.server.attributes?.logoGPS);
+  const logo = useSelector((state) => state.session.server.attributes?.logo);
   const logoInverted = useSelector((state) => state.session.server.attributes?.logoInverted);
 
   if (logo) {
