@@ -2,28 +2,31 @@ import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import makeStyles from '@mui/styles/makeStyles';
 import {
-  IconButton, Tooltip, Avatar, ListItemAvatar, ListItemText, ListItemButton,
+  IconButton, Tooltip, Avatar, ListItemAvatar, ListItemText, ListItemButton, 
 } from '@mui/material';
-import BuildCircleIcon from '@mui/icons-material/BuildCircle';
+import BatteryFullIcon from '@mui/icons-material/BatteryFull';
+import BatteryChargingFullIcon from '@mui/icons-material/BatteryChargingFull';
+import Battery60Icon from '@mui/icons-material/Battery60';
+import BatteryCharging60Icon from '@mui/icons-material/BatteryCharging60';
+import Battery20Icon from '@mui/icons-material/Battery20';
+import BatteryCharging20Icon from '@mui/icons-material/BatteryCharging20';
 import ErrorIcon from '@mui/icons-material/Error';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 import { devicesActions } from '../store';
 import {
-  formatAlarm, formatBoolean, formatPercentage, formatStatus, getStatusColor,
+  formatAlarm, formatBoolean, formatPercentage, formatStatus, getStatusColor, 
 } from '../common/util/formatter';
 import { useTranslation } from '../common/components/LocalizationProvider';
 import { mapIconKey, mapIcons } from '../map/core/preloadImages';
 import { useAdministrator } from '../common/util/permissions';
 import EngineIcon from '../resources/images/data/engine.svg?react';
 import { useAttributePreference } from '../common/util/preferences';
+import GppMaybeIcon from '@mui/icons-material/GppMaybe';
 
 dayjs.extend(relativeTime);
 
 const useStyles = makeStyles((theme) => ({
-  modelText: {
-    color: '#000000'
-  },
   icon: {
     width: '25px',
     height: '25px',
@@ -53,17 +56,16 @@ const DeviceRow = ({ data, index, style }) => {
   const dispatch = useDispatch();
   const t = useTranslation();
 
+  const server = useSelector((state) => state.session.server);
   const admin = useAdministrator();
 
   const item = data[index];
-  // console.log("🚀 ~ DeviceRow ~ item:", item)
   const position = useSelector((state) => state.session.positions[item.id]);
-
-  const events = useSelector((state) => state.events.items.filter((e) => e.deviceId === item.id));
-  const eventid = events[0];
 
   const devicePrimary = useAttributePreference('devicePrimary', 'name');
   const deviceSecondary = useAttributePreference('deviceSecondary', '');
+
+  const serverDaysoffline = server?.attributes?.daysoffline;
 
   const secondaryText = () => {
     let status;
@@ -75,20 +77,32 @@ const DeviceRow = ({ data, index, style }) => {
     return (
       <>
         {deviceSecondary && item[deviceSecondary] && `${item[deviceSecondary]} • `}
-        <span className={classes[getStatusColor(item.status, item.lastUpdate)]}>{status}</span>
+        <span className={classes[getStatusColor(item.status)]}>{status}</span>
       </>
     );
+  };
+
+  function daysfromNow(xDate) {
+    const currentDate = new Date();
+    const pastDate = new Date(xDate);
+    const timeDifference = currentDate - pastDate;
+    const daysDifference = Math.floor(timeDifference / (1000 * 60 * 60 * 24));
+    return daysDifference; // Return  number of days
   };
 
   const getStatusColorIcon = (item) => {
     switch (item.status) {
       case 'online':
       case 'unknown':
-        return '#22C55E';
+        if (daysfromNow(item.lastUpdate) >= serverDaysoffline) {
+          return 'red';
+        } else {
+          return 'green';
+        }
       case 'offline':
-        return '#EF4444';
+        return 'red';
       default:
-        return '#999999';
+        return 'gray';
     }
   };
 
@@ -106,9 +120,9 @@ const DeviceRow = ({ data, index, style }) => {
         </ListItemAvatar>
         <ListItemText
           primary={item[devicePrimary]}
-          primaryTypographyProps={{ sx: { color: 'black' }, noWrap: true }}
+          primaryTypographyProps={{ noWrap: true }}
           secondary={secondaryText()}
-          secondaryTypographyProps={{ sx: { color: '#999' }, noWrap: true }}
+          secondaryTypographyProps={{ noWrap: true }}
         />
         {position && (
           <>
@@ -119,37 +133,25 @@ const DeviceRow = ({ data, index, style }) => {
                 </IconButton>
               </Tooltip>
             )}
-            {/**Add Events Maintenance */}
-            {eventid && (
-              <>
-                {(eventid.hasOwnProperty('type') && eventid.type === "maintenance" && eventid.deviceId === item.id) && (
-                  <Tooltip title={`${eventid.attributes.message}`}>
-                    <IconButton size="small">
-                      <BuildCircleIcon fontSize="medium" className={classes.warning} />
-                    </IconButton>
-                  </Tooltip>
-                )}
-              </>
-            )}
             {position.attributes.hasOwnProperty('ignition') && (
-              <Tooltip title={position.attributes.output === 1 ? (t('commandEngineStop')) : (`${t('positionIgnition')}: ${formatBoolean(position.attributes.ignition, t)}`)}>
+              <Tooltip title={position.attributes.output === 1 ? (t('commandEngineStop')) : (`${t('positionIgnition')}: ${formatBoolean(position.attributes.ignition, t)}`) }>
                 <IconButton size="small">
                   {position.attributes.ignition ? (
-                    <EngineIcon width={25} height={25} className={position.attributes.output === 1 ? classes.error : classes.success} />
+                    <EngineIcon width={25} height={25} className={position.attributes.output === 1? classes.error : classes.success} />
                   ) : (
-                    <EngineIcon width={20} height={20} className={position.attributes.output === 1 ? classes.error : classes.neutral} />
+                    <EngineIcon width={20} height={20} className={position.attributes.output === 1? classes.error : classes.neutral} />
                   )}
                 </IconButton>
               </Tooltip>
             )}
-            {position.attributes.hasOwnProperty('ignition') && position.attributes.output === 1 && position.attributes.ignition && (
+            {position.attributes.hasOwnProperty('ignition') && position.attributes.output === 1 && position.attributes.ignition && (  
               <Tooltip title={`${t('alarmViolation')}`}>
                 <IconButton size="small">
-                  <ErrorIcon fontSize="small" className={classes.error} />
+                  <GppMaybeIcon fontSize="small" className={classes.error} />
                 </IconButton>
               </Tooltip>
             )}
-            {/*position.attributes.hasOwnProperty('batteryLevel') && (
+            {position.attributes.hasOwnProperty('batteryLevel') && (
               <Tooltip title={`${t('positionBatteryLevel')}: ${formatPercentage(position.attributes.batteryLevel)}`}>
                 <IconButton size="small">
                   {(position.attributes.batteryLevel > 70 && (
@@ -167,7 +169,7 @@ const DeviceRow = ({ data, index, style }) => {
                   )}
                 </IconButton>
               </Tooltip>
-            )*/}
+            )}
           </>
         )}
       </ListItemButton>
