@@ -5,7 +5,9 @@ import { useTheme } from '@mui/styles';
 import { map } from './core/MapView';
 import { formatTime, getStatusColor } from '../common/util/formatter';
 import { mapIconKey } from './core/preloadImages';
-import { useAttributePreference, usePreference } from '../common/util/preferences';
+import { useAttributePreference } from '../common/util/preferences';
+import { useCatchCallback } from '../reactHelper';
+import { findFonts } from './core/mapUtil';
 
 const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleField }) => {
   const id = useId();
@@ -20,7 +22,6 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
   const selectedDeviceId = useSelector((state) => state.devices.selectedId);
 
   const mapCluster = useAttributePreference('mapCluster', true);
-  const hours12 = usePreference('twelveHourFormat');
   const directionType = useAttributePreference('mapDirection', 'selected');
 
   const createFeature = (devices, position, selectedPositionId) => {
@@ -31,33 +32,22 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
         showDirection = false;
         break;
       case 'all':
-        showDirection = true;
+        showDirection = position.course > 0;
         break;
       default:
-        showDirection = selectedPositionId === position.id;
+        showDirection = selectedPositionId === position.id && position.course > 0;
         break;
     }
     return {
       id: position.id,
       deviceId: position.deviceId,
       name: device.name,
-      fixTime: formatTime(position.fixTime, 'seconds', hours12),
+      fixTime: formatTime(position.fixTime, 'seconds'),
       category: mapIconKey(device.category),
-      color: iconColor(device.status , position.attributes.color , position.attributes.ignition ),
+      color: showStatus ? position.attributes.color || getStatusColor(device.status) : 'neutral',
       rotation: position.course,
       direction: showDirection,
     };
-  };
-
-  const iconColor = ( status, attrcolor , ignition ) => {
-    let colors = showStatus ? attrcolor || getStatusColor(status) : 'neutral';
-    if( status == 'online' && !ignition ) {
-      colors = 'neutral';
-    }
-    else if( status == 'unknown' && ignition ) {
-      colors = 'success'; // support Cartrack data
-    }
-    return colors;
   };
 
   const onMouseEnter = () => map.getCanvas().style.cursor = 'pointer';
@@ -65,7 +55,7 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
 
   const onMapClick = useCallback((event) => {
     if (!event.defaultPrevented && onClick) {
-      onClick();
+      onClick(event.lngLat.lat, event.lngLat.lng);
     }
   }, [onClick]);
 
@@ -77,19 +67,16 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
     }
   }, [onClick]);
 
-  const onClusterClick = useCallback((event) => {
+  const onClusterClick = useCatchCallback(async (event) => {
     event.preventDefault();
     const features = map.queryRenderedFeatures(event.point, {
       layers: [clusters],
     });
     const clusterId = features[0].properties.cluster_id;
-    map.getSource(id).getClusterExpansionZoom(clusterId, (error, zoom) => {
-      if (!error) {
-        map.easeTo({
-          center: features[0].geometry.coordinates,
-          zoom,
-        });
-      }
+    const zoom = await map.getSource(id).getClusterExpansionZoom(clusterId);
+    map.easeTo({
+      center: features[0].geometry.coordinates,
+      zoom,
     });
   }, [clusters]);
 
@@ -103,7 +90,6 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
       cluster: mapCluster,
       clusterMaxZoom: 14,
       clusterRadius: 50,
-      
     });
     map.addSource(selected, {
       type: 'geojson',
@@ -126,6 +112,7 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
           'text-allow-overlap': true,
           'text-anchor': 'bottom',
           'text-offset': [0, -2 * iconScale],
+          'text-font': findFonts(map),
           'text-size': 12,
         },
         paint: {
@@ -164,6 +151,7 @@ const MapPositions = ({ positions, onClick, showStatus, selectedPosition, titleF
         'icon-image': 'background',
         'icon-size': iconScale,
         'text-field': '{point_count_abbreviated}',
+        'text-font': findFonts(map),
         'text-size': 14,
       },
     });
