@@ -2,8 +2,9 @@ import { useTheme } from '@mui/styles';
 import { useId, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { map } from './core/MapView';
+import getSpeedColor from '../common/util/colors';
 
-const MapRoutePath = ({ name, positions, coordinates }) => {
+const MapRoutePath = ({ positions }) => {
   const id = useId();
 
   const theme = useTheme();
@@ -19,7 +20,7 @@ const MapRoutePath = ({ name, positions, coordinates }) => {
         }
       }
     }
-    return theme.palette.geometry.main;
+    return null;
   });
 
   useEffect(() => {
@@ -46,21 +47,6 @@ const MapRoutePath = ({ name, positions, coordinates }) => {
         'line-width': 2,
       },
     });
-    if (name) {
-      map.addLayer({
-        source: id,
-        id: `${id}-title`,
-        type: 'symbol',
-        layout: {
-          'text-field': '{name}',
-          'text-size': 12,
-        },
-        paint: {
-          'text-halo-color': 'white',
-          'text-halo-width': 1,
-        },
-      });
-    }
 
     return () => {
       if (map.getLayer(`${id}-title`)) {
@@ -76,21 +62,30 @@ const MapRoutePath = ({ name, positions, coordinates }) => {
   }, []);
 
   useEffect(() => {
-    if (!coordinates) {
-      coordinates = positions.map((item) => [item.longitude, item.latitude]);
+    const minSpeed = positions.map((p) => p.speed).reduce((a, b) => Math.min(a, b), Infinity);
+    const maxSpeed = positions.map((p) => p.speed).reduce((a, b) => Math.max(a, b), -Infinity);
+    const features = [];
+    for (let i = 0; i < positions.length - 1; i += 1) {
+      features.push({
+        type: 'Feature',
+        geometry: {
+          type: 'LineString',
+          coordinates: [[positions[i].longitude, positions[i].latitude], [positions[i + 1].longitude, positions[i + 1].latitude]],
+        },
+        properties: {
+          color: reportColor || getSpeedColor(
+            positions[i + 1].speed,
+            minSpeed,
+            maxSpeed,
+          ),
+        },
+      });
     }
     map.getSource(id)?.setData({
-      type: 'Feature',
-      geometry: {
-        type: 'LineString',
-        coordinates,
-      },
-      properties: {
-        name,
-        color: reportColor,
-      },
+      type: 'FeatureCollection',
+      features,
     });
-  }, [theme, positions, coordinates, reportColor]);
+  }, [theme, positions, reportColor]);
 
   return null;
 };
