@@ -86,4 +86,10 @@ export const geofenceToFeature = (theme, item) => {
 
 export const geometryToArea = (geometry) => stringify(reverseCoordinates(geometry));
 
-
+export const findFonts = (map) => {
+  const { glyphs } = map.getStyle();
+  if (glyphs.startsWith('https://tiles.openfreemap.org')) {
+    return ['Noto Sans Regular'];
+  }
+  return ['Open Sans Regular', 'Arial Unicode MS Regular'];
+};
