@@ -1,11 +1,10 @@
 import dayjs from 'dayjs';
 import React, { useState } from 'react';
-import { useSelector } from 'react-redux';
 import {
-  FormControl, InputLabel, Select, MenuItem,
+  FormControl, InputLabel, Select, MenuItem, useTheme,
 } from '@mui/material';
 import {
-  AreaChart, Area, ResponsiveContainer, Tooltip, XAxis, YAxis,
+  Brush, CartesianGrid, Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAxis,
 } from 'recharts';
 import ReportFilter from './components/ReportFilter';
 import { formatTime } from '../common/util/formatter';
@@ -14,7 +13,7 @@ import PageLayout from '../common/components/PageLayout';
 import ReportsMenu from './components/ReportsMenu';
 import usePositionAttributes from '../common/attributes/usePositionAttributes';
 import { useCatch } from '../reactHelper';
-import { useAttributePreference, usePreference } from '../common/util/preferences';
+import { useAttributePreference } from '../common/util/preferences';
 import {
   altitudeFromMeters, distanceFromMeters, speedFromKnots, volumeFromLiters,
 } from '../common/util/converter';
@@ -22,28 +21,25 @@ import useReportStyles from './common/useReportStyles';
 
 const ChartReportPage = () => {
   const classes = useReportStyles();
+  const theme = useTheme();
   const t = useTranslation();
 
   const positionAttributes = usePositionAttributes(t);
-
-  const server = useSelector((state) => state.session.server);
-  const serverDarkMode = server?.attributes?.darkMode;
 
   const distanceUnit = useAttributePreference('distanceUnit');
   const altitudeUnit = useAttributePreference('altitudeUnit');
   const speedUnit = useAttributePreference('speedUnit');
   const volumeUnit = useAttributePreference('volumeUnit');
-  const hours12 = usePreference('twelveHourFormat');
 
   const [items, setItems] = useState([]);
-  const [types, setTypes] = useState(['fuel']);
-  const [type, setType] = useState('fuel');
+  const [types, setTypes] = useState(['speed']);
+  const [type, setType] = useState('speed');
+  const [timeType, setTimeType] = useState('fixTime');
 
   const values = items.map((it) => it[type]);
   const minValue = Math.min(...values);
   const maxValue = Math.max(...values);
   const valueRange = maxValue - minValue;
-
 
   const handleSubmit = useCatch(async ({ deviceId, from, to }) => {
     const query = new URLSearchParams({ deviceId, from, to });
@@ -58,6 +54,8 @@ const ChartReportPage = () => {
         const data = { ...position, ...position.attributes };
         const formatted = {};
         formatted.fixTime = dayjs(position.fixTime).valueOf();
+        formatted.deviceTime = dayjs(position.deviceTime).valueOf();
+        formatted.serverTime = dayjs(position.serverTime).valueOf();
         Object.keys(data).filter((key) => !['id', 'deviceId'].includes(key)).forEach((key) => {
           const value = data[key];
           if (typeof value === 'number') {
@@ -77,7 +75,7 @@ const ChartReportPage = () => {
                 formatted[key] = volumeFromLiters(value, volumeUnit).toFixed(2);
                 break;
               case 'hours':
-                formatted[key] = (value / 3600000).toFixed(2);
+                formatted[key] = (value / 1000).toFixed(2);
                 break;
               default:
                 formatted[key] = value;
@@ -100,20 +98,6 @@ const ChartReportPage = () => {
     }
   });
 
-  const colors = (serverDarkMode ? "#1F2A40" : "#e0e0e0")
-
-  const CustomTooltip = ({ active, payload, label }) => {
-    if (active && payload && payload.length) {
-      return (
-        <div className="custom-tooltip">
-          <p className="label">{`${formatTime(label, 'seconds', hours12)}`}</p>
-          <p className="label">{type} : {`${payload[0].value}`}</p>
-        </div>
-      );
-    }
-    return null;
-  };
-
   return (
     <PageLayout menu={<ReportsMenu />} breadcrumbs={['reportTitle', 'reportChart']}>
       <ReportFilter handleSubmit={handleSubmit} showOnly>
@@ -125,7 +109,6 @@ const ChartReportPage = () => {
               value={type}
               onChange={(e) => setType(e.target.value)}
               disabled={!items.length}
-              className={classes.selectFieldStyle}
             >
               {types.map((key) => (
                 <MenuItem key={key} value={key}>{positionAttributes[key]?.name || key}</MenuItem>
@@ -133,38 +116,65 @@ const ChartReportPage = () => {
             </Select>
           </FormControl>
         </div>
+        <div className={classes.filterItem}>
+          <FormControl fullWidth>
+            <InputLabel>{t('reportTimeType')}</InputLabel>
+            <Select
+              label={t('reportTimeType')}
+              value={timeType}
+              onChange={(e) => setTimeType(e.target.value)}
+              disabled={!items.length}
+            >
+              <MenuItem value="fixTime">{t('positionFixTime')}</MenuItem>
+              <MenuItem value="deviceTime">{t('positionDeviceTime')}</MenuItem>
+              <MenuItem value="serverTime">{t('positionServerTime')}</MenuItem>
+            </Select>
+          </FormControl>
+        </div>
       </ReportFilter>
       {items.length > 0 && (
         <div className={classes.chart}>
           <ResponsiveContainer>
-            <AreaChart
+            <LineChart
               data={items}
               margin={{
-                top: 10, right: 20, left: 15, bottom: 15,
+                top: 10, right: 40, left: 0, bottom: 10,
               }}
             >
-              <defs>
-                <linearGradient id="colorPv" x1="0" y1="0" x2="0" y2="1">
-                  <stop offset="5%" stopColor="#82ca9d" stopOpacity={0.8} />
-                  <stop offset="95%" stopColor="#82ca9d" stopOpacity={0} />
-                </linearGradient>
-              </defs>
               <XAxis
-                dataKey="fixTime"
+                stroke={theme.palette.text.primary}
+                dataKey={timeType}
                 type="number"
-                tickFormatter={(value) => formatTime(value, 'time', hours12)}
+                tickFormatter={(value) => formatTime(value, 'time')}
                 domain={['dataMin', 'dataMax']}
                 scale="time"
               />
               <YAxis
+                stroke={theme.palette.text.primary}
                 type="number"
                 tickFormatter={(value) => value.toFixed(2)}
-                domain={type === 'fuel' ? [0, 100] : [minValue - valueRange / 5 < 0 ? 0 : minValue - valueRange / 5, maxValue + valueRange / 5]}
+                domain={[minValue - valueRange / 5, maxValue + valueRange / 5]}
               />
-              {/*<CartesianGrid strokeDasharray="3 3" />*/}
-              <Tooltip content={<CustomTooltip />} />
-              <Area type="monotone" dataKey={type} stroke="#3da58a" fillOpacity={1} fill="url(#colorPv)" dot={false} connectNulls={true} />
-            </AreaChart>
+              <CartesianGrid stroke={theme.palette.divider} strokeDasharray="3 3" />
+              <Tooltip
+                contentStyle={{ backgroundColor: theme.palette.background.default, color: theme.palette.text.primary }}
+                formatter={(value, key) => [value, positionAttributes[key]?.name || key]}
+                labelFormatter={(value) => formatTime(value, 'seconds')}
+              />
+              <Brush
+                dataKey={timeType}
+                height={30}
+                stroke={theme.palette.primary.main}
+                tickFormatter={() => ''}
+              />
+              <Line
+                type="monotone"
+                dataKey={type}
+                stroke={theme.palette.primary.main}
+                dot={false}
+                activeDot={{ r: 6 }}
+              />
+            </LineChart>
           </ResponsiveContainer>
         </div>
       )}
