@@ -13,11 +13,6 @@ import TableShimmer from '../common/components/TableShimmer';
 import RemoveDialog from '../common/components/RemoveDialog';
 
 const useStyles = makeStyles((theme) => ({
-  fontStyle: {
-    '& .MuiTableCell-root': {
-      color: '#FFF'
-    }
-  },
   columnAction: {
     width: '1%',
     paddingRight: theme.spacing(1),
@@ -68,7 +63,7 @@ const ScheduledPage = () => {
 
   return (
     <PageLayout menu={<ReportsMenu />} breadcrumbs={['settingsTitle', 'reportScheduled']}>
-      <Table className={classes.fontStyle}>
+      <Table>
         <TableHead>
           <TableRow>
             <TableCell>{t('sharedType')}</TableCell>
