@@ -1,14 +1,14 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useRef  } from 'react';
 import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import {
-  FormControl, InputLabel, Select, MenuItem, Table, TableHead, TableRow, TableBody, TableCell, IconButton,
+  FormControl, InputLabel, Select, MenuItem, Table, TableHead, TableRow, TableBody, TableCell,
 } from '@mui/material';
 import {
-  formatDistance, formatSpeed, formatVolume, formatTime, formatNumericHours,
+  formatDistance, formatSpeed, formatVolume, formatTime, formatNumericHours, formatPercentage,
 } from '../common/util/formatter';
 import ReportFilter from './components/ReportFilter';
-import { useAttributePreference, usePreference } from '../common/util/preferences';
+import { useAttributePreference } from '../common/util/preferences';
 import { useTranslation } from '../common/components/LocalizationProvider';
 import PageLayout from '../common/components/PageLayout';
 import ReportsMenu from './components/ReportsMenu';
@@ -18,9 +18,6 @@ import { useCatch } from '../reactHelper';
 import useReportStyles from './common/useReportStyles';
 import TableShimmer from '../common/components/TableShimmer';
 import scheduleReport from './common/scheduleReport';
-import { useDownloadExcel } from 'react-export-table-to-excel';
-import ArrowCircleDownIcon from '@mui/icons-material/ArrowCircleDown';
-import Tooltip from '@mui/material/Tooltip';
 
 const columnsArray = [
   ['startTime', 'reportStartDate'],
@@ -30,6 +27,8 @@ const columnsArray = [
   ['averageSpeed', 'reportAverageSpeed'],
   ['maxSpeed', 'reportMaximumSpeed'],
   ['engineHours', 'reportEngineHours'],
+  ['startHours', 'reportStartEngineHours'],
+  ['endHours', 'reportEndEngineHours'],
   ['spentFuel', 'reportSpentFuel'],
 ];
 const columnsMap = new Map(columnsArray);
@@ -44,13 +43,11 @@ const SummaryReportPage = () => {
   const distanceUnit = useAttributePreference('distanceUnit');
   const speedUnit = useAttributePreference('speedUnit');
   const volumeUnit = useAttributePreference('volumeUnit');
-  const hours12 = usePreference('twelveHourFormat');
 
-  const [columns, setColumns] = usePersistedState('summaryColumns', ['startTime', 'engineHours', 'spentFuel']);
+  const [columns, setColumns] = usePersistedState('summaryColumns', ['startTime', 'distance', 'averageSpeed']);
   const [daily, setDaily] = useState(false);
   const [items, setItems] = useState([]);
   const [loading, setLoading] = useState(false);
-
 
   const handleSubmit = useCatch(async ({ deviceIds, groupIds, from, to, type }) => {
     const query = new URLSearchParams({ from, to, daily });
@@ -92,50 +89,40 @@ const SummaryReportPage = () => {
   });
 
   const formatValue = (item, key) => {
+    const value = item[key];
     switch (key) {
       case 'deviceId':
-        return devices[item[key]].name;
+        return devices[value].name;
       case 'startTime':
-      case 'endTime':
-        return formatTime(item[key], 'date', hours12);
+        return formatTime(value, 'date');
       case 'startOdometer':
       case 'endOdometer':
       case 'distance':
-        return formatDistance(item[key], distanceUnit, t);
+        return formatDistance(value, distanceUnit, t);
       case 'averageSpeed':
       case 'maxSpeed':
-        return formatSpeed(item[key], speedUnit, t);
+        return value > 0 ? formatSpeed(value, speedUnit, t) : null;
       case 'engineHours':
-        return formatNumericHours(item[key], t);
-      //const hours = Math.floor(item[key] / 3600000);
-      //const minutes = Math.floor((item[key] % 3600000) / 60000);
-      //return `${hours}:${minutes}`;
+      case 'startHours':
+      case 'endHours':
+        return value > 0 ? formatNumericHours(value, t) : null;
       case 'spentFuel':
-        return item[key].toFixed(1); //formatVolume(item[key], volumeUnit, t);
+        return value > 0 ? formatPercentage(value) : null;
       default:
-        return item[key];
+        return value;
     }
   };
 
   const tableRef = useRef(null);
 
-  const { onDownload } = useDownloadExcel({
-    currentTableRef: tableRef.current,
-    filename: 'SummaryReport',
-    sheet: 'SummaryReport'
-  });
-
-
   return (
     <PageLayout menu={<ReportsMenu />} breadcrumbs={['reportTitle', 'reportSummary']}>
       <div className={classes.header}>
-        <ReportFilter handleSubmit={handleSubmit} handleSchedule={handleSchedule} multiDevice includeGroups>
+        <ReportFilter handleSubmit={handleSubmit} handleSchedule={handleSchedule} multiDevice includeGroups loading={loading}>
           <div className={classes.filterItem}>
             <FormControl fullWidth>
               <InputLabel>{t('sharedType')}</InputLabel>
-              <Select label={t('sharedType')}
-                className={classes.selectFieldStyle}
-                value={daily} onChange={(e) => setDaily(e.target.value)}>
+              <Select label={t('sharedType')} value={daily} onChange={(e) => setDaily(e.target.value)}>
                 <MenuItem value={false}>{t('reportSummary')}</MenuItem>
                 <MenuItem value>{t('reportDaily')}</MenuItem>
               </Select>
@@ -144,16 +131,9 @@ const SummaryReportPage = () => {
           <ColumnSelect columns={columns} setColumns={setColumns} columnsArray={columnsArray} />
         </ReportFilter>
       </div>
-      <Table ref={tableRef} stickyHeader aria-label="sticky table" className={classes.tableStyle}>
+      <Table ref={tableRef} stickyHeader aria-label="sticky table">
         <TableHead>
           <TableRow>
-            <TableCell className={classes.columnAction}>
-              <Tooltip title="Download to excel">
-                <IconButton size="small" color="info" aria-label="download to excel" onClick={onDownload}>
-                  <ArrowCircleDownIcon />
-                </IconButton>
-              </Tooltip>
-            </TableCell>
             <TableCell>{t('sharedDevice')}</TableCell>
             {columns.map((key) => (<TableCell key={key}>{t(columnsMap.get(key))}</TableCell>))}
           </TableRow>
@@ -161,7 +141,6 @@ const SummaryReportPage = () => {
         <TableBody>
           {!loading ? items.map((item) => (
             <TableRow key={(`${item.deviceId}_${Date.parse(item.startTime)}`)}>
-              <TableCell />
               <TableCell>{devices[item.deviceId].name}</TableCell>
               {columns.map((key) => (
                 <TableCell key={key}>
