@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  FormControl, InputLabel, MenuItem, Select, useMediaQuery, useTheme
+  FormControl, InputLabel, MenuItem, Select,
 } from '@mui/material';
 import { useTranslation } from '../../common/components/LocalizationProvider';
 import useReportStyles from '../common/useReportStyles';
@@ -8,23 +8,19 @@ import useReportStyles from '../common/useReportStyles';
 const ColumnSelect = ({
   columns, setColumns, columnsArray, rawValues, disabled,
 }) => {
-  const theme = useTheme();
-  const isMediumOrSmaller = useMediaQuery(theme.breakpoints.down('md'));
   const classes = useReportStyles();
   const t = useTranslation();
-
 
   return (
     <div className={classes.filterItem}>
       <FormControl fullWidth>
-        {isMediumOrSmaller && <InputLabel>{t('sharedColumns')}</InputLabel>}
+        <InputLabel>{t('sharedColumns')}</InputLabel>
         <Select
           label={t('sharedColumns')}
           value={columns}
           onChange={(e) => setColumns(e.target.value)}
           multiple
           disabled={disabled}
-          className={classes.selectFieldStyle}
         >
           {columnsArray.map(([key, string]) => (
             <MenuItem key={key} value={key}>{rawValues ? string : t(string)}</MenuItem>
