@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import {
-  Table, TableRow, TableCell, TableHead, TableBody, Button, TableFooter,
+  Table, TableRow, TableCell, TableHead, TableBody, Button, TableFooter, FormControlLabel, Switch,
 } from '@mui/material';
 import LinkIcon from '@mui/icons-material/Link';
 import { useEffectAsync } from '../reactHelper';
@@ -13,10 +13,11 @@ import CollectionFab from './components/CollectionFab';
 import CollectionActions from './components/CollectionActions';
 import TableShimmer from '../common/components/TableShimmer';
 import SearchHeader, { filterByKeyword } from './components/SearchHeader';
-import { usePreference } from '../common/util/preferences';
 import { formatTime } from '../common/util/formatter';
-import { useDeviceReadonly } from '../common/util/permissions';
+import { useDeviceReadonly, useManager } from '../common/util/permissions';
 import useSettingsStyles from './common/useSettingsStyles';
+import DeviceUsersValue from './components/DeviceUsersValue';
+import usePersistedState from '../common/util/usePersistedState';
 
 const DevicesPage = () => {
   const classes = useSettingsStyles();
@@ -25,19 +26,20 @@ const DevicesPage = () => {
 
   const groups = useSelector((state) => state.groups.items);
 
-  const hours12 = usePreference('twelveHourFormat');
-
+  const manager = useManager();
   const deviceReadonly = useDeviceReadonly();
 
   const [timestamp, setTimestamp] = useState(Date.now());
   const [items, setItems] = useState([]);
   const [searchKeyword, setSearchKeyword] = useState('');
+  const [showAll, setShowAll] = usePersistedState('showAllDevices', false);
   const [loading, setLoading] = useState(false);
 
   useEffectAsync(async () => {
     setLoading(true);
     try {
-      const response = await fetch('/api/devices');
+      const query = new URLSearchParams({ all: showAll });
+      const response = await fetch(`/api/devices?${query.toString()}`);
       if (response.ok) {
         setItems(await response.json());
       } else {
@@ -46,7 +48,7 @@ const DevicesPage = () => {
     } finally {
       setLoading(false);
     }
-  }, [timestamp]);
+  }, [timestamp, showAll]);
 
   const handleExport = () => {
     window.location.assign('/api/reports/devices/xlsx');
@@ -72,6 +74,7 @@ const DevicesPage = () => {
             <TableCell>{t('deviceModel')}</TableCell>
             <TableCell>{t('deviceContact')}</TableCell>
             <TableCell>{t('userExpirationTime')}</TableCell>
+            {manager && <TableCell>{t('settingsUsers')}</TableCell>}
             <TableCell className={classes.columnAction} />
           </TableRow>
         </TableHead>
@@ -84,7 +87,8 @@ const DevicesPage = () => {
               <TableCell>{item.phone}</TableCell>
               <TableCell>{item.model}</TableCell>
               <TableCell>{item.contact}</TableCell>
-              <TableCell>{formatTime(item.expirationTime, 'date', hours12)}</TableCell>
+              <TableCell>{formatTime(item.expirationTime, 'date')}</TableCell>
+              {manager && <TableCell><DeviceUsersValue deviceId={item.id} /></TableCell>}
               <TableCell className={classes.columnAction} padding="none">
                 <CollectionActions
                   itemId={item.id}
@@ -96,12 +100,26 @@ const DevicesPage = () => {
                 />
               </TableCell>
             </TableRow>
-          )) : (<TableShimmer columns={7} endAction />)}
+          )) : (<TableShimmer columns={manager ? 8 : 7} endAction />)}
         </TableBody>
         <TableFooter>
           <TableRow>
-            <TableCell colSpan={8} align="right">
+            <TableCell>
               <Button onClick={handleExport} variant="text">{t('reportExport')}</Button>
+            </TableCell>
+            <TableCell colSpan={manager ? 8 : 7} align="right">
+              <FormControlLabel
+                control={(
+                  <Switch
+                    checked={showAll}
+                    onChange={(e) => setShowAll(e.target.checked)}
+                    size="small"
+                  />
+                )}
+                label={t('notificationAlways')}
+                labelPlacement="start"
+                disabled={!manager}
+              />
             </TableCell>
           </TableRow>
         </TableFooter>
