@@ -9,6 +9,7 @@ import {
   Checkbox,
   FormGroup,
   Button,
+  TextField,
 } from '@mui/material';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { useTranslation, useTranslationKeys } from '../common/components/LocalizationProvider';
@@ -93,7 +94,7 @@ const NotificationPage = () => {
               {item.notificators?.includes('command') && (
                 <SelectField
                   value={item.commandId}
-                  onChange={(event) => setItem({ ...item, commandId: Number(event.target.value) })}
+                  onChange={(e) => setItem({ ...item, commandId: Number(e.target.value) })}
                   endpoint="/api/commands"
                   titleGetter={(it) => it.description}
                   label={t('sharedSavedCommand')}
@@ -112,7 +113,7 @@ const NotificationPage = () => {
                   control={(
                     <Checkbox
                       checked={item.always}
-                      onChange={(event) => setItem({ ...item, always: event.target.checked })}
+                      onChange={(e) => setItem({ ...item, always: e.target.checked })}
                     />
                     )}
                   label={t('notificationAlways')}
@@ -127,9 +128,14 @@ const NotificationPage = () => {
               </Typography>
             </AccordionSummary>
             <AccordionDetails className={classes.details}>
+              <TextField
+                value={item.description || ''}
+                onChange={(e) => setItem({ ...item, description: e.target.value })}
+                label={t('sharedDescription')}
+              />
               <SelectField
                 value={item.calendarId}
-                onChange={(event) => setItem({ ...item, calendarId: Number(event.target.value) })}
+                onChange={(e) => setItem({ ...item, calendarId: Number(e.target.value) })}
                 endpoint="/api/calendars"
                 label={t('sharedCalendar')}
               />
