@@ -24,9 +24,11 @@ const CollectionFab = ({ editPath, disabled }) => {
 
   if (!readonly && !disabled) {
     return (
-      <Fab size="medium" color="primary" className={classes.fab} onClick={() => navigate(editPath)}>
-        <AddIcon />
-      </Fab>
+      <div className={classes.fab}>
+        <Fab size="medium" color="primary" onClick={() => navigate(editPath)}>
+          <AddIcon />
+        </Fab>
+      </div>
     );
   }
   return '';
