@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  Divider, List, ListItemButton, ListItemIcon, ListItemText,
+  Divider, List,
 } from '@mui/material';
 import SettingsIcon from '@mui/icons-material/Settings';
 import CreateIcon from '@mui/icons-material/Create';
@@ -15,22 +15,14 @@ import PublishIcon from '@mui/icons-material/Publish';
 import SmartphoneIcon from '@mui/icons-material/Smartphone';
 import HelpIcon from '@mui/icons-material/Help';
 import CampaignIcon from '@mui/icons-material/Campaign';
-import { Link, useLocation } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { useTranslation } from '../../common/components/LocalizationProvider';
 import {
   useAdministrator, useManager, useRestriction,
 } from '../../common/util/permissions';
 import useFeatures from '../../common/util/useFeatures';
-
-const MenuItem = ({
-  title, link, icon, selected,
-}) => (
-  <ListItemButton key={link} component={Link} to={link} selected={selected}>
-    <ListItemIcon>{icon}</ListItemIcon>
-    <ListItemText primary={title} />
-  </ListItemButton>
-);
+import MenuItem from '../../common/components/MenuItem';
 
 const SettingsMenu = () => {
   const t = useTranslation();
@@ -56,6 +48,12 @@ const SettingsMenu = () => {
         {!readonly && (
           <>
             <MenuItem
+              title={t('sharedNotifications')}
+              link="/settings/notifications"
+              icon={<NotificationsIcon />}
+              selected={location.pathname.startsWith('/settings/notification')}
+            />
+            <MenuItem
               title={t('settingsUser')}
               link={`/settings/user/${userId}`}
               icon={<PersonIcon />}
@@ -66,12 +64,6 @@ const SettingsMenu = () => {
               link="/settings/devices"
               icon={<SmartphoneIcon />}
               selected={location.pathname.startsWith('/settings/device')}
-            />
-            <MenuItem
-              title={t('sharedNotifications')}
-              link="/settings/notifications"
-              icon={<NotificationsIcon />}
-              selected={location.pathname.startsWith('/settings/notification')}
             />
             <MenuItem
               title={t('sharedGeofences')}
@@ -141,21 +133,19 @@ const SettingsMenu = () => {
         <>
           <Divider />
           <List>
+            <MenuItem
+              title={t('serverAnnouncement')}
+              link="/settings/announcement"
+              icon={<CampaignIcon />}
+              selected={location.pathname === '/settings/announcement'}
+            />
             {admin && (
-              <>
-                <MenuItem
-                  title={t('serverAnnouncement')}
-                  link="/settings/announcement"
-                  icon={<CampaignIcon />}
-                  selected={location.pathname === '/settings/announcement'}
-                />
-                <MenuItem
-                  title={t('settingsServer')}
-                  link="/settings/server"
-                  icon={<StorageIcon />}
-                  selected={location.pathname === '/settings/server'}
-                />
-              </>
+              <MenuItem
+                title={t('settingsServer')}
+                link="/settings/server"
+                icon={<StorageIcon />}
+                selected={location.pathname === '/settings/server'}
+              />
             )}
             <MenuItem
               title={t('settingsUsers')}
