@@ -6,32 +6,24 @@ import { VitePWA } from 'vite-plugin-pwa';
 /* eslint-disable no-template-curly-in-string */
 export default defineConfig(() => ({
   server: {
-    // port: 3000,
+    port: 3000,
     proxy: {
-      '/api/socket': 'ws://ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com:8082', //localhost , ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com
-      '/api': 'http://ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com:8082',
-    }, 
-  },
-  build: {
-    outDir: '../modern/dist',
-    rollupOptions: {
-      output: {
-        assetFileNames: 'assets/[name].[hash].[ext]'
-      }
-    }
-  },
-  resolve: {
-    alias: {
-      '@': '/src',
+      '/api/socket': 'ws://ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com:8082', //'ws://localhost:8082', ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com
+      '/api': 'http://ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com:8082', //'http://localhost:8082', ec2-18-143-114-161.ap-southeast-1.compute.amazonaws.com
     },
   },
-  assetsInclude: ['**/*.png', '**/*.jpg', '**/*.jpeg', '**/*.gif', '**/*.svg'],
+  build: {
+    outDir: 'build',
+  },
   plugins: [
     svgr(),
     react(),
     VitePWA({
+      includeAssets: ['favicon.ico', 'apple-touch-icon-180x180.png'],
       workbox: {
         navigateFallbackDenylist: [/^\/api/],
+        maximumFileSizeToCacheInBytes: 10 * 1024 * 1024,
+        globPatterns: ['**/*.{js,css,html,woff,woff2,mp3}'],
       },
       manifest: {
         short_name: '${title}',
